Use `invalid_onion_blinding` everywhere
Allowing any node inside the blinded route to return helpful errors to the
sender creates a probing attack vector for malicious senders, even for
errors coming from the final recipient, because it lets the sender know
that it was able to reach the recipient.

We now require every node in blinded paths to exclusively use
`invalid_onion_blinding`, and also convert downstream errors to
`invalid_onion_blinding` to ensure that buggy/malicious downstream nodes
are unable to exfiltrate data.